Property panel
I started working on the functionality for the property panel.
It opens up when the user right clicks on an object. it closes with a right click as well.

It has three categories:
- scale / size
- colour
- content (editing the text for: title, attributes, methods)

It's simple and working so far, but can be improved. Especially the way in which the text editor is opened