Buttongroup, more camera movement, script fixes
This update is a bit lengthy. I just forgot to check Github desktop to push some of this earlier :skull:

Buttongroup:
I wanted to change the current buttons at the bottom of the screen, to make them into radiobuttons. This would be helpful, as only one button in a group can be toggled. And so you (hopefully) don't get any design or logic errors. (hopefully)

I created a few style resources the buttons as well. (Normal, hover, pressed).

Right click + Camera movement:
- I changed the collision layer of the grid to be different to the spawned objects.
So when you right click on the grid and hold, then you can move around.
But if you right click on an object, then the property-panel opens up.