menuButton Items are working
I finally found out how to use the menuButton properly.
I add the menu items in code, dynamically,
and I also connect the popup for each menubutton, to a function.

I figured out that I had to use this notation: menuButton.get_popup(),id_pressed.connect(functionName)

and my initial method heading was not compatible for some reason. So i just renamed to "selectItem" with an integer param as Inded