New class shape + more properties manager
I made a new class shape. (Just duplicated it/ saved as a new file).
This one has three components (title, fields, and methods) which can be customized (size, and text displayed)

I also started connecting a few buttons to signals via code, dynamically, instead of assigning signals in the editor, which could get messy.
This means that the functions that i built, for binding the signals to buttons, are a bit longer but work really well :smile:

Also added small script to actor, so that title can be changed